fix: respect module boundary, move project-utils to datamodel/common
